mingw: special-case spawnve()ing MSYS2 programs
MSYS2 has funny quoting rules... So we have to accommodate for it by
quoting specifically when we detect that the executable is MSYS2 (for
which we use the tell-tale that the .exe lives in usr/bin/).

As the MSYS2 quoting rules are different from the regular Win32 ones,
and as we already handle the regular Win32 case, we now need to be
careful to choose either one or the other.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>