mingw: access all paths via \\?\
For historical reasons, Windows APIs restrict paths to a relatively
short length (260 characters or less).

There is a work-around, though, by using a special Win32 namespace:
prefix the path with \\?\ and access the APIs via the UTF-16 functions.

The only thing to be careful about is that \\?\ side-steps normalization
and validation of paths, i.e. the slashes must be backslashes, no . or
.. components are allowed, no runs of multiple slashes, etc.

This treatment also brings about a slightly unexpected performance
benefit, as the path validation and normalization is a bit more costly
than one would imagine: not only does the path validation have to
exclude components such as NUL, AUX, PRT1, etc, it also needs to make
sure that if any component is in 8.3 format, then the "8 part" cannot
carry any of those reserved names.

For the moment, this simply pretends that readlink() is not implemented
(we will need to override it eventually, anyway, what with Windows 10
Creators Update allowing user mode symlinks in Developer Mode).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>